moved the user role map to be in utils updated the emails and frontend to use all the same mapings
